requirements.txt
flask
google-generativeai
gTTS
python-dotenv